+++ roomserver/state/v1/state_test.go
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package state
+package v1
 
 import (
 	"testing"
